Interceptor, Login with Session and Action Error

+++ demo-struts2/src/main/java/com/example/basic/action/AccountAction.java
@@ -5,27 +5,30 @@ import com.example.basic.model.AccountModel;
 import com.example.basic.model.MessageStore;
 import com.example.basic.service.IAccountService;
 import com.example.basic.service.impl.AccountServiceImpl;
+import com.opensymphony.xwork2.ActionContext;
 import com.opensymphony.xwork2.ActionSupport;
 import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
+import org.apache.struts2.action.ParametersAware;
+import org.apache.struts2.action.SessionAware;
+import org.apache.struts2.dispatcher.HttpParameters;
 
 import java.util.List;
+import java.util.Map;
 
-public class AccountAction extends ActionSupport {
+public class AccountAction extends BaseAction {
+    @Getter
+    @Setter
     private MessageStore messageStore;
-
     @Getter
     @Setter
     private AccountModel accountModel;
-
     @Getter
     private List<Account> accounts;
-
     @Getter
     @Setter
     private int userId;
-
     @Getter
     @Setter
     private String usernameQuery;
@@ -37,7 +40,13 @@ public class AccountAction extends ActionSupport {
     }
 
     public String execute() {
-        messageStore = new MessageStore();
+        String actionError = getActionErrorFromSession();
+        addActionError(actionError);
+
+        String username = (String) ActionContext.getContext().getSession().get("username");
+        if (username == null) {
+            return LOGIN;
+        }
         accounts = accountService.selectAll();
         return SUCCESS;
     }
@@ -48,21 +57,24 @@ public class AccountAction extends ActionSupport {
     }
 
     public String deleteAccount() {
-        accountService.delete(userId);
+        int sessionUserId = (int) ActionContext.getContext().getSession().get("userId");
+        if (sessionUserId == userId) {
+            addActionErrorToSession("You cannot delete your own account");
+        } else {
+            accountService.delete(userId);
+        }
+
         return SUCCESS;
     }
 
     public String restoreAccount() {
         accountService.restore(userId);
+        accounts = accountService.selectAll();
         return SUCCESS;
     }
 
     public String searchAccount() {
         accounts = accountService.search(usernameQuery);
         return SUCCESS;
-    }
-
-    public MessageStore getMessageStore() {
-        return messageStore;
     }
 }
